[QoL] moved GateType from instruction.py to gameplay_config.py, so it can be accessed easier from other modules
