+++ qrogue/game/logic/collectibles/__init__.py
@@ -1,7 +1,7 @@
 # exporting
 from .collectible import Collectible, CollectibleType, MultiCollectible
 from .collectible_factory import CollectibleFactory, GateFactory, OrderedCollectibleFactory
-from .instruction import GateType, Instruction, InstructionManager
+from .instruction import Instruction, InstructionManager
 from .pickup import Pickup, Score, Key, Energy, Qubit
 
 # importing
